sync/requirements.txt
ynab==1.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
tenacity==8.2.3